added Edit Subject page  and changed  Context

+++ ElectronicDiaryApi/ModelsDto/UsersView/GroupDto.cs
@@ -7,11 +7,12 @@
         public string Classroom { get; set; }
         public LocationDto Location { get; set; }
         public List<EmployeeDto> Teachers { get; set; } = new();
-        public int MaxStudentCount { get; set; }  // StudentCount из модели
+        public sbyte MaxStudentCount { get; set; }  // StudentCount из модели
         public int CurrentStudents { get; set; } // Количество IdStudents
         public int IdSubject { get; set; }
         public string SubjectName { get; set; }
         public string MinAge { get; set; } = null!;
         public string MaxAge { get; set; } = null!;
+        //public bool? IsDelete { get; set; }
     }
 }
